+++ app/models/user_survey.rb
@@ -7,4 +7,12 @@ class UserSurvey < ApplicationRecord
 
   # Relationships
   belongs_to :user
+
+  def update_cache!(event_design)
+    update(
+      design_name_cache: event_design.design_name,
+      design_slug_cache: event_design.design_slug,
+      event_slug_cache: event_design.event_slug
+    )
+  end
 end