+++ db/migrate/20180509202609_add_event_design_cache_to_user_survey.rb
@@ -1,0 +1,7 @@
+class AddEventDesignCacheToUserSurvey < ActiveRecord::Migration[5.2]
+  def change
+    add_column :user_surveys, :design_name_cache, :string
+    add_column :user_surveys, :design_slug_cache, :string
+    add_column :user_surveys, :event_slug_cache, :string
+  end
+end